clients/go: go vet: fix packet user_data pointer semantics

+++ src/clients/go/tb_client.go
@@ -185,7 +185,7 @@ func (c *c_client) doRequest(
 
 	// Setup the packet.
 	req.packet.next = nil
-	req.packet.user_data = C.uintptr_t((uintptr)((unsafe.Pointer)(req)))
+	req.packet.user_data = unsafe.Pointer(req)
 	req.packet.operation = C.uint8_t(op)
 	req.packet.status = C.TB_PACKET_OK
 	req.packet.data_size = C.uint32_t(count * int(getEventSize(op)))
@@ -237,7 +237,7 @@ func onGoPacketCompletion(
 	result_len C.uint32_t,
 ) {
 	// Get the request from the packet user data
-	req := (*request)((unsafe.Pointer)((uintptr)(packet.user_data)))
+	req := (*request)(unsafe.Pointer(packet.user_data))
 	op := C.TB_OPERATION(packet.operation)
 
 	var wrote C.uint32_t
